Рефакторинг всего кода.

+++ src/lzwhash/lzwdefs.h
@@ -1,13 +1,15 @@
 #ifndef LZWDEFS_H
 #define LZWDEFS_H
 
+#include <cstdlib>
+
 namespace llzz {
 
 struct paramLZ {
-    int ntable;
-    int nsymbols;
-    int ncode;
-    int size_code_buff;
+    size_t ntable;
+    size_t nsymbols;
+    size_t ncode;
+    size_t size_code_buff;
 };
 
 }
